refactor: Power request messages
