+++ usbpd/src/protocol_layer/message/request.rs
@@ -1,11 +1,14 @@
+//! Definitions of request message content.
 use byteorder::{ByteOrder, LittleEndian};
 use proc_bitfield::bitfield;
-use uom::si::electric_current::centiampere;
+use uom::si::electric_current::{self, centiampere};
+use uom::si::u16::{ElectricCurrent, ElectricPotential};
 use uom::si::{self};
 
 use super::_20millivolts_mod::_20millivolts;
 use super::_250milliwatts_mod::_250milliwatts;
 use super::_50milliamperes_mod::_50milliamperes;
+use super::pdo;
 
 bitfield! {
     #[derive(Clone, Copy, PartialEq, Eq)]
@@ -162,7 +165,7 @@ impl Avs {
 }
 
 /// Power requests towards the source.
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Copy)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 #[allow(unused)] // FIXME: Implement missing request types.
 pub enum PowerSource {
@@ -171,6 +174,34 @@ pub enum PowerSource {
     Pps(Pps),
     Avs(Avs),
     Unknown(RawDataObject),
+}
+
+/// Errors that can occur during sink requests towards the source.
+#[derive(Debug)]
+#[non_exhaustive]
+pub enum Error {
+    /// A requested (specific) voltage does not exist in the PDOs.
+    VoltageMismatch,
+}
+
+/// Requestable voltage levels.
+#[derive(Debug)]
+pub enum VoltageRequest {
+    /// The safe 5 V supply.
+    Safe5V,
+    /// The highest voltage that the source can supply.
+    Highest,
+    /// A specific voltage.
+    Specific(ElectricPotential),
+}
+
+/// Requestable currents.
+#[derive(Debug)]
+pub enum CurrentRequest {
+    /// The highest current that the source can supply.
+    Highest,
+    /// A specific current.
+    Specific(ElectricCurrent),
 }
 
 impl PowerSource {
@@ -182,5 +213,93 @@ impl PowerSource {
             PowerSource::Avs(p) => p.object_position(),
             PowerSource::Unknown(p) => p.object_position(),
         }
+    }
+
+    /// Find the highest fixed voltage that can be found in the source capabilities.
+    ///
+    /// Reports the index of the found PDO, and the fixed supply instance, or `None` if there is no fixed supply PDO.
+    fn find_highest_fixed_voltage(source_capabilities: &pdo::SourceCapabilities) -> Option<(usize, &pdo::FixedSupply)> {
+        let mut selected_pdo = None;
+
+        for (index, cap) in source_capabilities.pdos().iter().enumerate() {
+            if let pdo::PowerDataObject::FixedSupply(fixed_supply) = cap {
+                selected_pdo = match selected_pdo {
+                    None => Some((index, fixed_supply)),
+                    Some(x) => {
+                        if fixed_supply.voltage() > x.1.voltage() {
+                            Some((index, fixed_supply))
+                        } else {
+                            selected_pdo
+                        }
+                    }
+                };
+            }
+        }
+
+        selected_pdo
+    }
+
+    /// Find a specific fixed voltage within the source capabilities.
+    ///
+    /// Reports the index of the found PDO, and the fixed supply instance, or `None` if there is no match to the request.
+    fn find_specific_fixed_voltage(
+        source_capabilities: &pdo::SourceCapabilities,
+        voltage: ElectricPotential,
+    ) -> Option<(usize, &pdo::FixedSupply)> {
+        for (index, cap) in source_capabilities.pdos().iter().enumerate() {
+            if let pdo::PowerDataObject::FixedSupply(fixed_supply) = cap {
+                if fixed_supply.voltage() == voltage {
+                    return Some((index, fixed_supply));
+                }
+            }
+        }
+
+        None
+    }
+
+    /// Create a new power source request for a fixed supply.
+    ///
+    /// Finds a suitable PDO by evaluating the provided current and voltage requests against the source capabilities.
+    pub fn new_fixed(
+        current_request: CurrentRequest,
+        voltage_request: VoltageRequest,
+        source_capabilities: &pdo::SourceCapabilities,
+    ) -> Result<Self, Error> {
+        let selected = match voltage_request {
+            VoltageRequest::Safe5V => source_capabilities.vsafe_5v().map(|supply| (0, supply)),
+            VoltageRequest::Highest => Self::find_highest_fixed_voltage(source_capabilities),
+            VoltageRequest::Specific(x) => Self::find_specific_fixed_voltage(source_capabilities, x),
+        };
+
+        if selected.is_none() {
+            return Err(Error::VoltageMismatch);
+        }
+
+        let (index, supply) = selected.unwrap();
+
+        let (current, mismatch) = match current_request {
+            CurrentRequest::Highest => (supply.max_current(), false),
+            CurrentRequest::Specific(x) => (x, x > supply.max_current()),
+        };
+
+        let mut raw_current = current.get::<electric_current::centiampere>();
+
+        if raw_current > 0x3ff {
+            error!("Clamping invalid current: {} mA", 10 * raw_current);
+            raw_current = 0x3ff;
+        }
+
+        let object_position = index + 1;
+        assert!(object_position > 0b0000 && object_position <= 0b1110);
+
+        Ok(Self::FixedVariableSupply(
+            FixedVariableSupply(0)
+                .with_raw_operating_current(raw_current)
+                .with_raw_max_operating_current(raw_current)
+                .with_object_position(object_position as u8)
+                .with_capability_mismatch(mismatch)
+                .with_no_usb_suspend(true)
+                .with_usb_communications_capable(true), // FIXME: Make adjustable?
+        ))
     }
 }
