+++ usbpd/src/sink/mod.rs
@@ -2,4 +2,3 @@
 
 pub mod device_policy_manager;
 pub mod policy_engine;
-pub mod request;
